Replace HTML header code with Dynamic JS Header
Big change to how header is used on this website.

Instead of each page having its own HTML for the header, it is now
contained in resources/common/header.txt, which contains the HTML that
is included in a javascript call.  To change the header, now you only
have to change the header.txt file.